chore: renames constants file to shared
