+++ test/fields/collections/Collapsible/shared.ts
